requirements.txt
fastapi==0.110.0
uvicorn==0.30.1
pydantic==2.6.4
pycryptodome==3.20.0
argon2-cffi==23.1.0
colorama==0.4.6
pytest==8.1.1